requirements_upto_28_03_2021.txt
amqp==5.0.5
anyjson==0.3.3
arabic-reshaper==2.1.1
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.6
defusedxml==0.6.0
Django==3.0.4
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.12.4
future==0.18.2
httplib2==0.19.0
idna==2.10
importlib-metadata==3.7.0
kombu==5.0.2
Markdown==2.6.11
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.19.5
oauth2==1.9.0.post1
oauthlib==3.1.0
pandas==0.25.3
parler==1.0.1
paypal==1.2.5
prompt-toolkit==3.0.3
protobuf==3.15.2
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
python-social-auth==0.2.12
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
zipp==3.4.0
